src/requirements.txt
RPi.GPIO
ipython
google-assistant-library
google-assistant-sdk[samples]
google-auth-oauthlib[tool]
yaml